# backend/openapi.yaml
openapi: 3.1.0
info:
  title: THUIE API
  version: 1.0.0-draft
  description: >-
    REST + WebSocket contract for the THUIE campus alumni info-exchange
    platform. Design skeleton - mirrors BACKEND.md and the PostgreSQL schema in
    PROJECT.md 6. Server implementation deferred to the build phase.
  license: { name: Proprietary }
servers:
  - url: https://api.thuie.example.com/api/v1
    description: Production (placeholder)
  - url: http://localhost:3000/api/v1
    description: Local development

# OpenAPI has no native WebSocket binding, so the realtime surface (BACKEND.md §9)
# is documented here as a vendor extension. Socket.IO server, namespace
# `/realtime`, authenticated with the same access JWT supplied via handshake
# `auth.token`, an `Authorization: Bearer` header, or a `?token=` query.
...
security:
  - bearerAuth: []
paths:
...
  /infos:
    get: { tags: [Infos], operationId: listInfos, summary: Info feed (cursor), parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: category, schema: { $ref: '#/components/schemas/InfoCategory' } }, { $ref: '#/components/parameters/Sort' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/InfoPostPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Infos], operationId: createInfo, summary: Submit info (-> moderation), parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InfoCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/InfoPost' } } } }, '422': { description: keyword blocked, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /infos/{id}:
    get: { tags: [Infos], operationId: getInfo, summary: Single info, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/InfoPost' } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Infos], operationId: updateInfo, summary: Author edit (approved -> re-review), parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/InfoCreate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/InfoPost' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Infos], operationId: deleteInfo, summary: Author takedown, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: taken down }, default: { $ref: '#/components/responses/Error' } } }
  /infos/{id}/report:
    post: { tags: [Infos], operationId: reportInfo, summary: Report an info post, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ReportCreate' } } } }, responses: { '202': { description: filed }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Alumni ----
...
  /posts:
    get: { tags: [Forum], operationId: listPosts, summary: Forum feed (cursor), parameters: [ { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }, { in: query, name: tag, schema: { type: string } }, { in: query, name: q, schema: { type: string } } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPostPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Forum], operationId: createPost, summary: Submit post (-> moderation), parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/PostCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPost' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /posts/{id}:
    get: { tags: [Forum], operationId: getPost, summary: Post detail (view counted), parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPost' } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }
    patch: { tags: [Forum], operationId: updatePost, summary: Author edit, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/PostCreate' } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPost' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Forum], operationId: deletePost, summary: Author takedown, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: taken down }, default: { $ref: '#/components/responses/Error' } } }
  /posts/{id}/like:
    post: { tags: [Forum], operationId: toggleLike, summary: Toggle like, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPost' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /posts/{id}/bookmark:
    post: { tags: [Forum], operationId: toggleBookmark, summary: Toggle bookmark (favorites(post)), parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/ForumPost' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /posts/{id}/comments:
    get: { tags: [Forum], operationId: listComments, summary: Comment thread, parameters: [ { $ref: '#/components/parameters/Id' }, { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/CommentPage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Forum], operationId: createComment, summary: Add comment (keyword-pass -> approved), parameters: [ { $ref: '#/components/parameters/Id' }, { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/CommentCreate' } } } }, responses: { '201': { description: created, content: { application/json: { schema: { $ref: '#/components/schemas/Comment' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /posts/{id}/report:
    post: { tags: [Forum], operationId: reportPost, summary: Report a post, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ReportCreate' } } } }, responses: { '201': { description: filed }, default: { $ref: '#/components/responses/Error' } } }
  /comments/{id}:
    patch: { tags: [Forum], operationId: updateComment, summary: Edit comment (version-checked), parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { type: object, properties: { content: { type: string }, version: { type: integer } } } } } }, responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/Comment' } } } }, '412': { description: version conflict, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Forum], operationId: deleteComment, summary: Author takedown, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '204': { description: taken down }, default: { $ref: '#/components/responses/Error' } } }
  /comments/{id}/report:
    post: { tags: [Forum], operationId: reportComment, summary: Report a comment, parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/ReportCreate' } } } }, responses: { '201': { description: filed }, default: { $ref: '#/components/responses/Error' } } }
...
  /favorites:
    get: { tags: [Engagement], operationId: listFavorites, summary: Favorites by type (rows carry the target's title/subtitle), parameters: [ { in: query, name: type, schema: { $ref: '#/components/schemas/FavoriteTargetType' } }, { $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/FavoritePage' } } } }, default: { $ref: '#/components/responses/Error' } } }
    post: { tags: [Engagement], operationId: addFavorite, summary: Add favorite (idempotent), requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/FavoriteCreate' } } } }, responses: { '201': { description: ok }, default: { $ref: '#/components/responses/Error' } } }
    delete: { tags: [Engagement], operationId: removeFavorite, summary: "Remove favorite (same triple; /favorites/{type}/{id} alias)", requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/FavoriteCreate' } } } }, responses: { '200': { description: ok }, default: { $ref: '#/components/responses/Error' } } }
  /favorites/status:
    get: { tags: [Engagement], operationId: favoriteStatus, summary: Existence probe for one target (paints the save icon), parameters: [ { in: query, name: type, required: true, schema: { $ref: '#/components/schemas/FavoriteTargetType' } }, { in: query, name: id, required: true, schema: { type: string } } ], responses: { '200': { description: ok, content: { application/json: { schema: { type: object, properties: { favorited: { type: boolean } } } } } }, default: { $ref: '#/components/responses/Error' } } }
  /favorites/{type}/{id}:
    delete: { tags: [Engagement], operationId: removeFavoriteAlias, summary: Remove favorite (alias), parameters: [ { in: path, name: type, required: true, schema: { $ref: '#/components/schemas/FavoriteTargetType' } }, { in: path, name: id, required: true, schema: { type: string } } ], responses: { '200': { description: ok }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Notifications ----
...
  /media:
    post: { tags: [Media], operationId: requestMediaUpload, summary: Create media_object + presigned PUT, requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/MediaRequest' } } } }, responses: { '201': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/MediaUpload' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /media/{id}:
    get: { tags: [Media], operationId: getMedia, summary: Media metadata / signed GET, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/MediaObject' } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }
  /media/{id}/content:
    put: { tags: [Media], operationId: uploadMediaContent, summary: Upload bytes (local fallback for the presigned PUT), parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { required: true, content: { 'application/octet-stream': { schema: { type: string, format: binary } } } }, responses: { '200': { description: stored, content: { application/json: { schema: { $ref: '#/components/schemas/MediaObject' } } } }, default: { $ref: '#/components/responses/Error' } } }
  /media/{id}/thumb:
    get: { tags: [Media], security: [], operationId: getMediaThumb, summary: Thumbnail bytes (public, for <img>), parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: image bytes, content: { 'image/*': { schema: { type: string, format: binary } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }
  /media/{id}/{name}:
    get: { tags: [Media], security: [], operationId: getMediaFile, summary: File bytes (public, for <img>), parameters: [ { $ref: '#/components/parameters/Id' }, { in: path, name: name, required: true, schema: { type: string }, description: Ignored slug; the media id selects the object } ], responses: { '200': { description: file bytes, content: { 'application/octet-stream': { schema: { type: string, format: binary } } } }, '404': { $ref: '#/components/responses/NotFound' }, default: { $ref: '#/components/responses/Error' } } }

  # ---- Admin / moderation ----
...
  /reports:
    post: { tags: [Engagement], operationId: createReport, summary: File a report (any target), parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReportCreateGeneric' } } } }, responses: { '202': { description: filed }, default: { $ref: '#/components/responses/Error' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    Limit: { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
    Id: { in: path, name: id, required: true, schema: { type: string } }
...
    Cursor: { in: query, name: cursor, schema: { type: string }, description: Opaque keyset cursor }
    Sort: { in: query, name: sort, schema: { type: string }, description: "-created_at (leading - = DESC)" }
    IdempotencyKey: { in: header, name: Idempotency-Key, schema: { type: string, format: uuid } }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Not found (or hidden by visibility)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    Role: { type: string, enum: [student, graduate, admin, admin_super] }
    UserStatus: { type: string, enum: [unverified, active, posting_restricted, banned, deleted] }
...
    ContentStatus: { type: string, enum: [pending, approved, rejected, taken_down] }
...
    Visibility: { type: string, enum: [student_only, all, admin_only] }
    InfoCategory: { type: string, enum: [internal, open, recruitment] }
    InfoSource: { type: string, enum: [official, user] }
...
    FavoriteTargetType: { type: string, enum: [info, profile, post, comment] }
...
    ReportTargetType: { type: string, enum: [info_post, forum_post, comment, alumni_profile, user] }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable machine code }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }
        errors: { type: array, items: { type: object, properties: { field: { type: string }, message: { type: string } } } }
    CursorMeta: { type: object, properties: { nextCursor: { type: [string, 'null'] }, limit: { type: integer } } }
...
    Department: { type: object, properties: { id: { type: string }, code: { type: string }, name_zh: { type: string }, name_en: { type: string }, faculty: { type: [string, 'null'] } } }
    User: { type: object, properties: { id: { type: string }, name: { type: string }, role: { $ref: '#/components/schemas/Role' }, status: { $ref: '#/components/schemas/UserStatus' }, student_id: { type: [string, 'null'] }, phone: { type: [string, 'null'] }, email: { type: [string, 'null'] }, department: { $ref: '#/components/schemas/Department' }, grade_year: { type: [string, 'null'] }, graduation_year: { type: [string, 'null'] }, avatar_url: { type: [string, 'null'] }, bio: { type: [string, 'null'] }, wechat: { type: [string, 'null'], description: GET /me only }, whatsapp: { type: [string, 'null'], description: GET /me only }, linkedin: { type: [string, 'null'], description: GET /me only }, wechat_visibility: { $ref: '#/components/schemas/Visibility' }, whatsapp_visibility: { $ref: '#/components/schemas/Visibility' }, linkedin_visibility: { $ref: '#/components/schemas/Visibility' }, email_visibility: { $ref: '#/components/schemas/Visibility' }, phone_visibility: { $ref: '#/components/schemas/Visibility' }, is_online: { type: boolean }, created_at: { type: string, format: date-time } } }
...
    MediaObject: { type: object, properties: { id: { type: string }, kind: { type: string, enum: [image, attachment, avatar] }, url: { type: string }, file_name: { type: string }, mime_type: { type: string }, size_bytes: { type: integer }, width: { type: [integer, 'null'] }, height: { type: [integer, 'null'] }, thumbnail_key: { type: [string, 'null'] } } }
    MediaRequest: { type: object, required: [kind, file_name, mime_type, size_bytes], properties: { kind: { type: string, enum: [image, attachment, avatar] }, file_name: { type: string }, mime_type: { type: string }, size_bytes: { type: integer }, sha256: { type: string } } }
    MediaUpload: { type: object, properties: { media_id: { type: string }, upload_url: { type: string }, expires_at: { type: string, format: date-time } } }
    InfoPost: { type: object, properties: { id: { type: string }, title: { type: string }, content: { type: string }, category: { $ref: '#/components/schemas/InfoCategory' }, source: { $ref: '#/components/schemas/InfoSource' }, visibility: { $ref: '#/components/schemas/Visibility' }, status: { $ref: '#/components/schemas/ContentStatus' }, pinned: { type: boolean }, version: { type: integer }, author: { $ref: '#/components/schemas/User' }, media: { type: array, items: { $ref: '#/components/schemas/MediaObject' } }, created_at: { type: string, format: date-time }, edited_at: { type: [string, 'null'], format: date-time } } }
    InfoCreate: { type: object, required: [title, content], properties: { title: { type: string, maxLength: 200 }, content: { type: string }, category: { $ref: '#/components/schemas/InfoCategory' }, visibility: { $ref: '#/components/schemas/Visibility' }, media_ids: { type: array, items: { type: string } } } }
...
    Tag: { type: object, properties: { id: { type: string }, name: { type: string }, slug: { type: string } } }
    ForumPost: { type: object, properties: { id: { type: string }, title: { type: string }, content: { type: string }, author: { $ref: '#/components/schemas/User' }, status: { $ref: '#/components/schemas/ContentStatus' }, location: { type: [string, 'null'] }, is_pinned: { type: boolean }, tags: { type: array, items: { $ref: '#/components/schemas/Tag' } }, view_count: { type: integer }, like_count: { type: integer }, bookmark_count: { type: integer }, comment_count: { type: integer }, liked: { type: boolean }, bookmarked: { type: boolean }, liked_by: { type: array, maxItems: 3, items: { $ref: '#/components/schemas/User' } }, version: { type: integer }, created_at: { type: string, format: date-time }, last_comment_at: { type: [string, 'null'], format: date-time } } }
    PostCreate: { type: object, required: [title, content], properties: { title: { type: string }, content: { type: string }, location: { type: string }, tags: { type: array, items: { type: string } } } }
    Comment: { type: object, properties: { id: { type: string }, forum_post_id: { type: string }, parent_id: { type: [string, 'null'] }, author: { $ref: '#/components/schemas/User' }, content: { type: string }, status: { $ref: '#/components/schemas/ContentStatus' }, version: { type: integer }, created_at: { type: string, format: date-time } } }
    CommentCreate: { type: object, required: [content], properties: { content: { type: string }, parent_id: { type: [string, 'null'] } } }
...
    Favorite: { type: object, properties: { target_type: { $ref: '#/components/schemas/FavoriteTargetType' }, target_id: { type: string }, created_at: { type: string, format: date-time }, title: { type: [string, 'null'] }, subtitle: { type: [string, 'null'] } } }
    FavoriteCreate: { type: object, required: [target_type, target_id], properties: { target_type: { $ref: '#/components/schemas/FavoriteTargetType' }, target_id: { type: string } } }
...
    ReportCreate: { type: object, required: [reason], properties: { reason: { type: string, maxLength: 500 } } }
...
    ReportCreateGeneric: { type: object, required: [target_type, target_id, reason], properties: { target_type: { $ref: '#/components/schemas/ReportTargetType' }, target_id: { type: string }, reason: { type: string, maxLength: 500 } } }
...
    InfoPostPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/InfoPost' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
...
    ForumPostPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/ForumPost' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
    CommentPage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Comment' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
...
    FavoritePage: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Favorite' } }, meta: { $ref: '#/components/schemas/CursorMeta' } } }
...